If stream_field not configured, use schema_uri_field to create topic
This adds some extra default behavior if events don't have stream
or topic names in them.  This makes the explanation and usage
of a default eventbus install simpler; The 'only' extra information
clients need to include in their events is the schema URI.
If stream_field is not configued, Extracted schema URIs will be
sanitized and used as the Kafka topic.