Add context to Eventbus.process, validate and produce functions
This allows for users of Eventbus (like the service template route)
to pass in extra context for validate() and produce() implementaitons
to use.

This is used by wikimedia-eventbus to instantiate two different
Kafka producers, a guarunteed and a 'hasty' producer, to use
for both reliable and fire and forget ?hasty=true cases.