Use schema 'title' field as schema name in stream config
This allows us to use the title rather to ensure that an event
of a particular schema is allowed in the stream, rather than
relying on the schema_uri.  The title will be more consistent
and canonical than schema_uris, which might change due to location
moving, etc.

Also refactory wikimedia-eventbus a bit to simplify.