Support draft 4 schemas and schema_uri_field as array
The default Ajv instance knows about the Draft 7 JSONSchema meta schema,
but does not on its own include meta schemas for other JSONSchema drafts.
Schemas should point to their meta schema in the $schema field.
JSONSChema meta schemas are self validating, so they have their
$id fields and $schema fields set to the same URL.  If Ajv
does not yet have a meta schema referneced by that URL, this will
cause infinite recursion as our loadSchema implementation will
look up remote schemas by $id.

This patch adds a option for EventValidator to match $id schema URLs
as being meta schemas, instead of data (event) schemas.  If a schema's
URL matches, it will be added as a meta schema to Ajv, breaking the
infinite recursion.

---

This patch also includes a change to the default-eventgate schema_uri_field
option.  If this is given as an array, the event will be searched for
the field names in the array when extracting a schema uri.

Both of thse changes allow wikimedia to support draft 4 events
with schema URIs at meta.schema_uri, as well as new draft 7 ones
we will make with schema URIs at $schema.