Fix bug with runtime remote loading of meta schemas
The use of _.defaults in EventValidator constructor was
causing the defaultOptions.ajvConfig to be assigned
to the final value of ajvConfig.  By using defaultsDeep
only once, we avoid accidentally assinging it and
possibly accidentally using the wrong loadSchema function
for different EventValidator instances.

This brings back the ability to add meta schemas at runtime
via the metaSchemaIdRegex option.