Fix bug in /v1/_test/events where test event default values were only being set once
Downstream processors of events might mutate those events if they are filling
in default values. Usually this is fine since normal events are discarded after
they are processed.  But we were holding a reference to test_events, and when
they were mutated, the next call to /v1/_test/events was passing in the mutated
events from the previous call.