Make node-rdkafka an optionalDependency
Producing to Kafka is just one way to use EventGate.
It's nice to have it as the default, except that building the
librdkafka binary is not very lightway, nor does it always work so
smoothly on all OS-es (e.g. MacOS). By making node-rdkafka an
optional dependency (and guarding against it being missing),
we should be able to more easily create development eventgate
instances without needing to build librdkafka.

Bug: T266058