Test validation of "const" keyword
A constant is a syntastic sugar for an enum having a single item.
However the spec does not mention anything regarding the expected type.

I am assuming validation is done with a strict equality and there is
thus no need to explicitly set the type. After all if one expects
integer value 42, there is no need to first validate the type then the
value.

Thus if one has a schema with:

  {
    "type": "object",
      "properties": {
        "kind": {
          "const": "hello"
        },
        "val": {
          "const": 3,
        }
      }
  }

Then `kind` must be exactly the string "hello" and `val` must be the
integer 3 and there is no need to specify the type.

I asked for potential clarification of the spec:
https://github.com/json-schema-org/json-schema-spec/issues/1256

Which got an answer from Henry Andrews (one of the json schema spec
author):

> Each keyword operates independently unless the specification
> explicitly describes an interaction.
>
> const requires the exact value, so it's not possible to match the
> value and not match the type. But that doesn't meant that there's an
> implicit type keyword, or that type needs to be explicitly present.